🔨 refactor: categoryTopLayout -\> 서스펜스 감싸기 위해 CategoryContent.tsx
파일 생성 후 분리
